requirements.txt
###### Requirements without Version Specifiers ######
streamlit
psycopg2-binary
pandas
numpy
matplotlib